requirements.txt
streamlit
pandas
requests
plotly
python-dotenv
fmpsdk
pygwalker